[MODINVOICE-133] - Implement the "upload" endpoint Refactor logic
