+++ src/test/java/org/folio/rest/impl/ApiTestSuite.java
@@ -9,7 +9,7 @@ import org.folio.converters.BatchVoucherModelConverterTest;
 import org.folio.converters.BatchedVoucherLinesModelConverterTest;
 import org.folio.converters.BatchedVoucherModelConverterTest;
 import org.folio.invoices.events.handlers.InvoiceSummaryTest;
-import org.folio.invoices.util.FtpUploadServiceTest;
+import org.folio.services.ftp.FtpUploadServiceTest;
 import org.folio.invoices.util.HelperUtilsTest;
 import org.folio.jaxb.DefaultJAXBRootElementNameResolverTest;
 import org.folio.jaxb.JAXBUtilTest;
@@ -19,9 +19,9 @@ import org.folio.rest.impl.protection.InvoicesProtectionTest;
 import org.folio.rest.impl.protection.LinesProtectionTest;
 import org.folio.rest.tools.utils.NetworkUtils;
 import org.folio.schemas.xsd.BatchVoucherSchemaXSDTest;
-import org.folio.services.BatchVoucherGenerateServiceTest;
+import org.folio.services.voucher.BatchVoucherGenerateServiceTest;
 import org.folio.services.InvoiceRetrieveServiceTest;
-import org.folio.services.UploadBatchVoucherExportServiceTest;
+import org.folio.services.voucher.UploadBatchVoucherExportServiceTest;
 import org.folio.services.VoucherLinesRetrieveServiceTest;
 import org.folio.services.ftp.FTPVertxCommandLoggerTest;
 import org.junit.jupiter.api.AfterAll;
