+++ src/test/java/org/folio/services/ftp/FTPVertxCommandLoggerTest.java
@@ -1,7 +1,5 @@
 package org.folio.services.ftp;
 
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
