+++ src/test/java/org/folio/services/voucher/BatchVoucherGenerateServiceTest.java
@@ -1,4 +1,4 @@
-package org.folio.services;
+package org.folio.services.voucher;
 
 import static org.folio.invoices.utils.HelperUtils.OKAPI_URL;
 import static org.folio.rest.impl.ApiTestSuite.mockPort;
@@ -14,7 +14,6 @@ import java.util.concurrent.ExecutionException;
 import org.folio.rest.impl.ApiTestBase;
 import org.folio.rest.jaxrs.model.BatchVoucher;
 import org.folio.rest.jaxrs.model.BatchVoucherExport;
-import org.folio.services.voucher.BatchVoucherGenerateService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
